# src/main/resources/static/api.yaml
openapi: 3.0.1
info:
  title: SerieApi
  description: >-
    API REST para gerenciamento de séries, plataformas de streaming e usuários.
    Permite cadastrar, atualizar, filtrar e associar séries a plataformas de streaming.
  version: v1.0.0
  contact:
    name: Suporte SerieApi
    email: suporte@serieapi.com.br
security:
  - bearerAuth: [ ]
servers:
  - url: http://localhost:8080
    description: Ambiente de desenvolvimento local
...
paths:
...
  /series:
    get:
      tags:
        - Series
      summary: Lista todas as séries
      description: Retorna todas as séries cadastradas no banco de dados de forma irrestrita.
      operationId: listarSeries
      responses:
        '200':
          description: Séries encontradas com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SerieResponseDto'
              examples:
                Sucesso:
                  summary: Lista de séries retornada com sucesso
                  value:
                    - id: 1
                      titulo: Breaking Bad
                      totalTemporada: 5
                      avaliacao: 9.5
                      atores: Bryan Cranston, Aaron Paul
                      poster: https://link-da-imagem.com/breaking-bad.jpg
                      sinopse: Um professor de química se volta para o crime.
                      dataLancamento: 2008-01-20
                      ativo: true
                      categoriaId:
                        id: 1
                        categoria: Drama
                    - id: 2
                      titulo: Stranger Things
                      totalTemporada: 4
                      avaliacao: 8.7
                      atores: Winona Ryder, David Harbour
                      poster: https://link-da-imagem.com/stranger-things.jpg
                      sinopse: Crianças enfrentam forças sobrenaturais.
                      dataLancamento: 2016-07-15
                      ativo: true
                      categoriaId:
                        id: 2
                        categoria: Ficção Científica
        '500':
          description: Erro interno no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroInternoResponse'
              examples:
                ErroInterno:
                  summary: Erro inesperado no servidor
                  value:
                    type: https://serieapi.com.br/erro-interno
                    title: Erro interno no servidor
                    status: 500
                    detail: Ocorreu um erro inesperado. Tente novamente mais tarde.
                    instance: /series

    post:
      tags:
        - Series
      summary: Cria uma nova série
      description: Cria e persiste uma nova série no banco de dados com os dados informados.
      operationId: salvarSerie
      requestBody:
        description: Dados necessários para a criação da série
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SerieRequestDto'
            examples:
              ExemploRequest:
                summary: Exemplo de criação de série
                value:
                  titulo: Breaking Bad
                  totalTemporada: 5
                  avaliacao: 9.5
                  atores: Bryan Cranston, Aaron Paul
                  poster: https://link-da-imagem.com/breaking-bad.jpg
                  sinopse: Um professor de química se volta para o crime.
                  dataLancamento: 2008-01-20
                  ativo: true
                  categoriaId:
                    id: 1
      responses:
        '201':
          description: Série criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SerieResponseDto'
              examples:
                Sucesso:
                  summary: Série criada com sucesso
                  value:
                    id: 1
                    titulo: Breaking Bad
                    totalTemporada: 5
                    avaliacao: 9.5
                    atores: Bryan Cranston, Aaron Paul
                    poster: https://link-da-imagem.com/breaking-bad.jpg
                    sinopse: Um professor de química se volta para o crime.
                    dataLancamento: 2008-01-20
                    ativo: true
                    categoriaId:
                      id: 1
                      categoria: Drama
        '400':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroDadosInvalidosResponse'
              examples:
                CamposInvalidos:
                  summary: Campos obrigatórios não preenchidos ou inválidos
                  value:
                    type: https://serieapi.com.br/dados-invalidos
                    title: Dados inválidos
                    status: 400
                    detail: Um ou mais campos estão incorretos. Faça o preenchimento correto.
                    instance: /series
                    fields:
                      - fieldName: titulo
                        detail: Título da série é obrigatório.
                      - fieldName: totalTemporada
                        detail: Total de temporadas deve ser maior ou igual a zero.
                      - fieldName: avaliacao
                        detail: Avaliação deve ser maior ou igual a zero.
  /series/{serieId}:
    get:
      tags:
        - Series
      summary: Busca uma série pelo ID
      description: Retorna os dados completos de uma série específica pelo seu ID.
      operationId: buscarSeriePorId
      parameters:
        - $ref: '#/components/parameters/serieId'
      responses:
        '200':
          description: Série encontrada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SerieResponseDto'
              examples:
                Sucesso:
                  summary: Série retornada com sucesso
                  value:
                    id: 1
                    titulo: Breaking Bad
                    totalTemporada: 5
                    avaliacao: 9.5
                    atores: Bryan Cranston, Aaron Paul
                    poster: https://link-da-imagem.com/breaking-bad.jpg
                    sinopse: Um professor de química se volta para o crime.
                    dataLancamento: 2008-01-20
                    ativo: true
                    categoriaId:
                      id: 1
                      categoria: Drama
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'
              examples:
                NaoEncontrado:
                  summary: Série não encontrada com o ID informado
                  value:
                    type: https://serieapi.com.br/recurso-nao-encontrado
                    title: Recurso não encontrado
                    status: 404
                    detail: Não existe uma série cadastrada com o código 99.
                    instance: /series/99

    put:
      tags:
        - Series
      summary: Atualiza uma série pelo ID
      description: Atualiza completamente os dados de uma série existente com base no ID informado.
      operationId: atualizarSerie
      parameters:
        - $ref: '#/components/parameters/serieId'
      requestBody:
        description: Novos dados da série
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SerieRequestDto'
            examples:
              ExemploRequest:
                summary: Exemplo de atualização de série
                value:
                  titulo: Breaking Bad — Edição Especial
                  totalTemporada: 5
                  avaliacao: 9.8
                  atores: Bryan Cranston, Aaron Paul
                  poster: https://link-da-imagem.com/breaking-bad-v2.jpg
                  sinopse: Um professor de química se volta para o crime.
                  dataLancamento: 2008-01-20
                  ativo: true
                  categoriaId:
                    id: 1
      responses:
        '200':
          description: Série atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SerieResponseDto'
        '400':
          description: Dados de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroDadosInvalidosResponse'
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'

    delete:
      tags:
        - Series
      summary: Remove uma série pelo ID
      description: Remove permanentemente uma série do banco de dados com base no ID informado.
      operationId: excluirSerie
      parameters:
        - $ref: '#/components/parameters/serieId'
      responses:
        '204':
          description: Série removida com sucesso
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'
        '409':
          description: Série em uso e não pode ser removida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroConflitoResponse'
              examples:
                EmUso:
                  summary: Série associada a um streaming
                  value:
                    type: https://serieapi.com.br/entidade-em-uso
                    title: Entidade em uso
                    status: 409
                    detail: A série de código 1 não pode ser removida pois está em uso.
                    instance: /series/1
  /series/{serieId}/ativar:
    put:
      tags:
        - Series
      summary: Ativa uma série
      description: Altera o status da série para ativo, indicando que ainda recebe novos episódios.
      operationId: ativarSerie
      parameters:
        - $ref: '#/components/parameters/serieId'
      responses:
        '204':
          description: Série ativada com sucesso
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'
  /series/{serieId}/desativar:
    delete:
      tags:
        - Series
      summary: Desativa uma série
      description: Altera o status da série para inativo, indicando que não recebe mais novos episódios.
      operationId: desativarSerie
      parameters:
        - $ref: '#/components/parameters/serieId'
      responses:
        '204':
          description: Série desativada com sucesso
        '404':
          description: Série não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroNaoEncontradoResponse'

  #  /series/filtros:
  #    get:
  #      tags:
  #        - Series
  #      summary: Filtra séries dinamicamente
  #      description: Busca séries aplicando filtros opcionais combinados. Todos os parâmetros são opcionais.
  #      operationId: listarSeriesComFiltros
  #      parameters:
  #        - name: titulo
  #          in: query
  #          description: Filtra séries pelo título (busca parcial)
  #          required: false
  #          schema:
  #            type: string
  #          example: Breaking
  #        - name: notaMax
  #          in: query
  #          description: Filtra séries com avaliação até o valor informado
  #          required: false
  #          schema:
  #            type: number
  #            format: double
  #          example: 9.5
  #        - name: limite
  #          in: query
  #          description: Limita a quantidade de resultados retornados
  #          required: false
  #          schema:
  #            type: integer
  #            format: int32
  #          example: 10
  #      responses:
  #        '200':
  #          description: Filtro processado com sucesso
  #          content:
  #            application/json:
  #              schema:
  #                type: array
  #                items:
  #                  $ref: '#/components/schemas/SerieResponseDto'
  #        '500':
  #          description: Erro interno no servidor
  #          content:
  #            application/json:
  #              schema:
  #                $ref: '#/components/schemas/ErroInternoResponse'

  # ==========================================
  # SÉRIES x STREAMINGS
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    serieId:
      name: serieId
      in: path
      description: ID numérico da série
      required: true
      schema:
        type: integer
        format: int64
      example: 1
...
  schemas:
...
    SerieRequestDto:
      type: object
      description: Dados para criação ou atualização de uma série
      required:
        - titulo
        - totalTemporada
        - avaliacao
        - atores
        - poster
        - sinopse
        - dataLancamento
        - categoriaId
      properties:
        titulo:
          type: string
          minLength: 1
          maxLength: 100
          description: Título oficial da série
          example: Breaking Bad
        totalTemporada:
          type: integer
          format: int32
          minimum: 0
          description: Quantidade total de temporadas lançadas
          example: 5
        avaliacao:
          type: number
          format: double
          minimum: 0
          description: Nota de avaliação da série (0 a 10)
          example: 9.5
        atores:
          type: string
          minLength: 1
          description: Nomes dos principais atores do elenco
          example: Bryan Cranston, Aaron Paul
        poster:
          type: string
          minLength: 1
          description: URL do poster ou imagem de capa da série
          example: https://link-da-imagem.com/poster.jpg
        sinopse:
          type: string
          minLength: 1
          description: Sinopse ou resumo do enredo
          example: Um professor de química se volta para o crime...
        dataLancamento:
          type: string
          format: date
          description: Data do lançamento original no formato yyyy-MM-dd
          example: 2008-01-20
        ativo:
          type: boolean
          description: Indica se a série ainda recebe novos episódios
          example: true
        categoriaId:
          $ref: '#/components/schemas/CategoriaIdRequestDto'
...
    CategoriaIdRequestDto:
      type: object
      description: Referência ao ID de uma categoria já cadastrada
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: ID numérico da categoria já cadastrada no banco de dados
          example: 1
...
    SerieResponseDto:
      type: object
      description: Dados completos de uma série retornados pela API
      required:
        - atores
        - avaliacao
        - dataLancamento
        - poster
        - sinopse
        - titulo
        - totalTemporada
        - categoriaId
      properties:
        id:
          type: integer
          format: int64
          description: ID único da série
          example: 1
        titulo:
          type: string
          description: Título oficial da série
          example: Breaking Bad
        totalTemporada:
          type: integer
          format: int32
          description: Quantidade total de temporadas
          example: 5
        avaliacao:
          type: number
          format: double
          description: Nota de avaliação da série
          example: 9.5
        atores:
          type: string
          description: Principais atores do elenco
          example: Bryan Cranston, Aaron Paul
        poster:
          type: string
          description: URL do poster da série
          example: https://link-da-imagem.com/poster.jpg
        sinopse:
          type: string
          description: Resumo do enredo
          example: Um professor de química se volta para o crime...
        dataLancamento:
          type: string
          format: date
          description: Data do lançamento original
          x-field-extra-annotation: "@com.fasterxml.jackson.annotation.JsonFormat(pattern = \"yyyy-MM-dd\")"
        ativo:
          type: boolean
          description: Indica se a série ainda recebe novos episódios
          example: true
        categoriaId:
          $ref: '#/components/schemas/CategoriaResponseDto'
    CategoriaResponseDto:
      type: object
      description: Dados de uma categoria retornados pela API
      properties:
        id:
          type: integer
          format: int64
          example: 1
        categoria:
          type: string
          example: Drama
...
    ErroDadosInvalidosResponse:
      type: object
      description: Resposta de erro para dados inválidos (400)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/dados-invalidos
        title:
          type: string
          example: Dados inválidos
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Um ou mais campos estão incorretos. Faça o preenchimento correto.
        instance:
          type: string
          example: /series
        fields:
          type: array
          items:
            $ref: '#/components/schemas/CampoInvalidoResponse'
    CampoInvalidoResponse:
      type: object
      description: Detalhe de um campo inválido
      properties:
        fieldName:
          type: string
          example: titulo
        detail:
          type: string
          example: Título da série é obrigatório.
    ErroNaoEncontradoResponse:
      type: object
      description: Resposta de erro para recurso não encontrado (404)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/recurso-nao-encontrado
        title:
          type: string
          example: Recurso não encontrado
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Não existe uma série cadastrada com o código 99.
        instance:
          type: string
          example: /series/99
    ErroConflitoResponse:
      type: object
      description: Resposta de erro para conflito de dados (409)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/entidade-em-uso
        title:
          type: string
          example: Entidade em uso
        status:
          type: integer
          example: 409
        detail:
          type: string
          example: A série de código 1 não pode ser removida pois está em uso.
        instance:
          type: string
          example: /series/1
...
    ErroInternoResponse:
      type: object
      description: Resposta de erro para falha interna no servidor (500)
      properties:
        type:
          type: string
          example: https://serieapi.com.br/erro-interno
        title:
          type: string
          example: Erro interno no servidor
        status:
          type: integer
          example: 500
        detail:
          type: string
          example: Ocorreu um erro inesperado. Tente novamente mais tarde.
        instance:
          type: string
          example: /series